pyproject.toml
[tool.poetry]
name = "workmate-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
python-dotenv = "^1.0.1"
djoser = "^2.2.3"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
